requirements.txt
python-telegram-bot==20.0
google-genai
httpx
notion-client
python-dotenv
requests
